Refactor App structure by separating tutorial and comment routes into dedicated files; enhance tests with mock implementations for improved coverage.
